First Pass removal of game-specific code
